App content/requirements.txt
clipboard
cx_Oracle
keyboard
mysql-connector-python
numpy
opencv-python
openpyxl
pandas
Pillow
pyautogui
pygetwindow
pymsgbox
PyQt6
pyodbc
pywin32
selenium
